# [feature-wip](multi-catalog) implement predicate pushdown in native OrcReader